WaE in OSL_DEBUG_LEVEL > 1

+++ vcl/unx/generic/app/salinst.cxx
@@ -146,7 +146,7 @@ bool X11SalInstance::AnyInput(VclInputFlags nType)
         bRet = aInput.bRet;
     }
 #if OSL_DEBUG_LEVEL > 1
-    fprintf( stderr, "AnyInput 0x%x = %s\n", nType, bRet ? "true" : "false" );
+    fprintf( stderr, "AnyInput 0x%x = %s\n", static_cast<unsigned int>(nType), bRet ? "true" : "false" );
 #endif
     return bRet;
 }
